+++ src/webpack/webpack.config-distribution.js
@@ -12,7 +12,7 @@ module.exports = function (env) {
   );
 
   const resources = dpat.Resources.copyDescriptors(buildManifest, PROJECT_ROOT_PATH);
-  const bundlePackages = dpat.BuildUtils.bundlePackages(PROJECT_ROOT_PATH, 'devDependencies');
+  const bundlePackages = dpat.BuildUtils.bundlePackages(PROJECT_ROOT_PATH, 'devDependencies').concat(['deskpro-components']);
   const babelOptions = dpat.Babel.resolveOptions(PROJECT_ROOT_PATH, { babelrc: false });
   // the relative path of the assets inside the distribution bundle
   const ASSET_PATH = 'assets';
@@ -52,26 +52,13 @@ module.exports = function (env) {
           include: [ path.resolve(PROJECT_ROOT_PATH, 'src/main/sass') ],
           loader: extractCssPlugin.extract({ use: ['css-loader', 'sass-loader'] })
         },
-        {
-          test: /\.(png|jpg)$/,
-          use: 'url-loader?limit=15000'
-        },
-        {
-          test: /\.eot(\?v=\d+.\d+.\d+)?$/,
-          use: 'file-loader'
-        },
-        {
-          test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/,
-          use: 'url-loader?limit=10000&mimetype=application/font-woff'
-        },
-        {
-          test: /\.[ot]tf(\?v=\d+.\d+.\d+)?$/,
-          use: 'url-loader?limit=10000&mimetype=application/octet-stream'
-        },
-        {
-          test: /\.svg(\?v=\d+\.\d+\.\d+)?$/,
-          use: 'url-loader?limit=10000&mimetype=image/svg+xml'
-        }
+
+        { test: /\.(png|jpg)$/, loader: 'url-loader', options: { limit: 15000 } },
+        { test: /\.eot(\?v=\d+.\d+.\d+)?$/, loader: 'file-loader' },
+        { test: /\.[ot]tf(\?v=\d+.\d+.\d+)?$/, loader: 'url-loader', options: { limit: 10000, mimetype: 'application/octet-stream' } },
+        { test: /\.svg(\?v=\d+\.\d+\.\d+)?$/, loader: 'url-loader', options: { limit: 10000, mimetype: 'image/svg+xml' } },
+        { test: /\.woff(2)?(\?v=[0-9]\.[0-9]\.[0-9])?$/, loader: 'url-loader', options: { limit: 10000, mimetype: 'application/font-woff' } }
+
       ],
     },
     output: {