+++ CHANGELOG.md
@@ -4,6 +4,24 @@ This project is following [Semantic Versioning](http://semver.org)
 
 ## [Unreleased][]
 
+## [0.2.0][] - 2017-10-09
+
+### Fixed
+
+ - using instance id when creating custom fields
+
+### Changed
+
+ - upgrade dependencies
+ - use deskpro components 
+
+## [0.1.0][] - 2017-10-03
+
 ### Added
 
  - initial release
+
+
+
+[Unreleased]: https://github.com/DeskproApps/app-installer/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/DeskproApps/app-installer/compare/v0.1.0...v0.2.0